Remove unused PROF_COUNTER functions
The function to pop a dialog is not really part of PROF_COUNTER,
if wanted, it can be done by the called.

This removes WX as a dep for PROF_COUNTER and profile,
which now used std::chrono only.

Also remove the print-to-stderr-with-delta function, as it's
unused too, and the delta times can be collected directly now
as std::chrono::durations if needed.