pcbnew: Fix 45° snapping polygons
The constrained draw was constraining the polygon size to fit the
existing points rather than extending the existing points to fit the
user's intent.

This adds a 90° double-joint for the completion polygon that allows the
user to complete common constrained polygons visually based on the
preview with a double-click to finish.

(cherry picked from commit bae50fab9f26b1ea42f3b72df66097ffb0d705d1)

Fixes: lp:1849299
* https://bugs.launchpad.net/kicad/+bug/1849299