Guard FP_LIB_TABLE rows with a mutex
Each FP_LIB_TABLE_ROW holds an instance of a PLUGIN. PLUGINs
are not generally thread-safe, but by accessing the same row
of a FP_LIB_TABLE from different threads, concurrent access to
PLUGINs is possible.

Introduce a very simple RAII wrapper for protecting a pointer
with a mutex. The motivation is to make it harder to accidentally
use the pointer without locking it. Because of the friend-ness of the
ROWs, it's not a complete fix.

Additionally, the plugins can still leak pointers to cached footprints
via, e.g. GetEnumeratedFootprint(). So this patch alone will not
eliminate the threat of use-after-free when PLUGINs are concurrently
accessed.

Fixes: lp:1853380
* https://bugs.launchpad.net/kicad/+bug/1853380