QA: Add a library loading plugin to the pcbnew QA tools
This allows to load a number of libraries from the command line, which
can be used for a variety of purposes: debugging, profiling and testing
the library loading functions.

It can load libraries directly from disk, or it can do it via an
existing fp-lib-table file and use nicknames.

Add tests using this executable:

* One test to directly load libs from the libraries
* One test to load via a library table. Notably, this test case can
expose the bug reported in lp:1853380 by concurrently loading from the
same library PLUGIN.

This introduces some footprints into the qa/data/footprints directory,
which are used as test data. Taken from the kicad-footprints repo, these
are licenced CC-BY-SA-4.0.