Ask to save changes even after pressing "Cancel"
When pressing "Cancel" in Cvpcb's Footprint Assignment window, Cvpcb
discards changes without prompting the user. The comment suggests that
this behavior is intentional.

I have personally lost work due to this behavior when I tried to click
"Ok" and missed. Discarding changes without a prompt seems like a bad
idea anyway, so this stop doing that.

The only disadvantage is that it is now harder to open and close the
Footprint Assignment window without saving changes.