Readd text sketch mode to the footprint editor toolbars
This was removed during the transition to tool actions,
but the setting still exists, so it can get enabled with
no way to turn it off. Another solution would be to remove
the option for text sketch completely, but I am too lazy
to do that right now.

Fixes https://gitlab.com/kicad/code/kicad/issues/4269